# Keep track of position when deleting from a BTreeMap