This improves the performance of drain_filter and is needed for future Cursor support for BTreeMap.

cc @ssomers 
r? @Mark-Simulacrum