@Mark-Simulacrum 2020-04-05
Looks good to me, thanks!

We should really look into caching the last_leaf_edge/first_leaf_edge I suspect, to avoid the (presumably) fairly expensive tree traversals on every non-leaf node we're visiting.

@bors r+