@Mark-Simulacrum 2020-04-05
I think further improvements to the enum can be cleared up later, for now this seems fine. @bors r+